+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change `rate` cli arg to `interval`
+
 ## [0.4.3] - 2020-02-14
 
 ### Fixed