+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2020-02-16
+
 ### Fixed
 
 - [process] fix process CPU calculation
@@ -106,7 +108,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release!
 
-[Unreleased]: https://github.com/cjbassi/ytop/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/cjbassi/ytop/compare/0.5.1...HEAD
+[0.5.1]: https://github.com/cjbassi/ytop/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/cjbassi/ytop/compare/0.4.3...0.5.0
 [0.4.3]: https://github.com/cjbassi/ytop/compare/0.4.2...0.4.3
 [0.4.2]: https://github.com/cjbassi/ytop/compare/0.4.1...0.4.2